Renamed i18n utilities file to fix typo (#517)
