+++ src/main/webapp/resources/js/components/galaxy/GalaxySubmission.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useStateValue } from "./GalaxyState";
 import { Button } from "antd";
-import { getI18N } from "../../utilities/i18n-utilties";
+import { getI18N } from "../../utilities/i18n-utilities";
 import { actions } from "./reducer";
 import {
   getGalaxySamples,
