+++ src/main/webapp/resources/js/components/pipeline/Pipeline.jsx
@@ -5,7 +5,7 @@ import styled from "styled-components";
 import { FONT_SIZE_DEFAULT, FONT_WEIGHT_HEAVY } from "../../styles/fonts";
 import { SPACE_SM } from "../../styles/spacing";
 import { BREAK_MD_MAX, BREAK_XL_MAX } from "../../styles/break-points";
-import { getI18N } from "../../utilities/i18n-utilties";
+import { getI18N } from "../../utilities/i18n-utilities";
 
 const PipelineCard = styled(Card)`
   .ant-card-body {
