+++ src/main/webapp/resources/js/pages/analyses/AnalysesPage.jsx
@@ -3,7 +3,7 @@ import { render } from "react-dom";
 import { AnalysesTable } from "../../components/AnalysesTable/AnalysesTable";
 import { PagedTableProvider } from "../../contexts/PagedTableContext";
 import { PageWrapper } from "../../components/page/PageWrapper";
-import { getI18N } from "../../utilities/i18n-utilties";
+import { getI18N } from "../../utilities/i18n-utilities";
 
 render(
   <PageWrapper title={getI18N("analyses.header")}>
