+++ src/main/webapp/resources/js/utilities/i18n-utilities.js
